Move ActivityResult out of ActivityStreams
