+++ gnd/src/main/java/com/google/android/gnd/system/ActivityStreams.java
@@ -107,40 +107,6 @@ public class ActivityStreams {
   }
 
   /** Represents the arguments of an {@link Activity#onActivityResult(int, int, Intent)} event. */
-  public static class ActivityResult {
-    private final int requestCode;
-    private final int resultCode;
-    @Nullable private final Intent data;
-
-    public ActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-      this.requestCode = requestCode;
-      this.resultCode = resultCode;
-      this.data = data;
-    }
-
-    /** Returns the activity request code for which this result applies. */
-    public int getRequestCode() {
-      return requestCode;
-    }
-
-    /**
-     * Returns true iff the system provided a result of {@link Activity#RESULT_OK} to the {@code
-     * onActivityResult} callback.
-     */
-    public boolean isOk() {
-      return resultCode == Activity.RESULT_OK;
-    }
-
-    /**
-     * Returns {@link Intent} data provided by the system to the {@code onActivityResult} callback.
-     */
-    @Nullable
-    public Intent getData() {
-      return data;
-    }
-  }
-
-  /** Represents the arguments of an {@link Activity#onActivityResult(int, int, Intent)} event. */
   public static class RequestPermissionsResult {
     private final int requestCode;
     private final Map<String, Integer> permissionGrantResults;
