+++ gnd/src/main/java/com/google/android/gnd/system/auth/GoogleAuthenticationManager.java
@@ -29,8 +29,8 @@ import com.google.android.gms.tasks.Task;
 import com.google.android.gnd.R;
 import com.google.android.gnd.model.User;
 import com.google.android.gnd.rx.annotations.Hot;
+import com.google.android.gnd.system.ActivityResult;
 import com.google.android.gnd.system.ActivityStreams;
-import com.google.android.gnd.system.ActivityStreams.ActivityResult;
 import com.google.android.gnd.system.auth.SignInState.State;
 import com.google.firebase.auth.AuthCredential;
 import com.google.firebase.auth.AuthResult;
